+++ CHANGES.md
@@ -6,6 +6,7 @@ Changes in 12.6
 - HTML responsive stackable tables in Export.php
 - Export list: add Total row in ListOutput.fnc.php
 - Add rounded flag icons in locale/[locale_code]/flag.png
+- Fix #362 Order by Points as Letter Grade is alphabetically sorted (may be inaccurate) in Grades.php
 
 Changes in 12.5
 ---------------